+++ exclusion.h
@@ -13,36 +13,36 @@
 
 using namespace std;
 
-inline unsigned int letter_to_index(const char & c);
+inline size_t letter_to_index(const char & c);
 
-inline char index_to_letter(const unsigned int & i);
+inline char index_to_letter(const size_t & i);
 
 inline void write_letter_table(vector<bool> & table,
-        const unsigned int & n_words,
-        const unsigned int & letter_i, const unsigned int & word_i,
+        const size_t & n_words,
+        const size_t & letter_i, const size_t & word_i,
         const bool value);
 
 inline bool read_letter_table(const vector<bool> & table,
-        const unsigned int & n_words,
-        const unsigned int & letter_i, const unsigned int & word_i);
+        const size_t & n_words,
+        const size_t& letter_i, const size_t & word_i);
 
 vector<bool> make_letter_table(const vector<string> & words);
 
-vector<unsigned int> make_unused_letter_indices(const vector<bool> & table,
-        const unsigned int & n_words,
+vector<size_t> make_unused_letter_indices(const vector<bool> & table,
+        const size_t & n_words,
         const vector<char> & used_letters,
-        const unordered_set<unsigned int> & word_indices);
+        const unordered_set<size_t> & word_indices);
 
-unsigned int lower_bound_on_remaining_words(const vector<bool> & table,
-        const unsigned int & n_words,
-        const unordered_set<unsigned int> & word_indices,
-        const vector<unsigned int> & unused_letter_indices,
-        const unsigned int & lives);
+size_t lower_bound_on_remaining_words(const vector<bool> & table,
+        const size_t & n_words,
+        const unordered_set<size_t> & word_indices,
+        const vector<size_t> & unused_letter_indices,
+        const size_t & lives);
 
-unsigned int upper_bound_on_remaining_words(
-        const unordered_set<unsigned int> & word_indices,
-        const vector<unsigned int> & unused_letter_indices,
+size_t upper_bound_on_remaining_words(
+        const unordered_set<size_t> & word_indices,
+        const vector<size_t> & unused_letter_indices,
         const context_t & ctx,
-        const unsigned int & lives);
+        const size_t & lives);
 
 #endif /* EXCLUSION_H */
